pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2.3",
    "plotly-express>=0.4.1",
    "plotly>=6.0.0",
    "requests>=2.32.3",
    "streamlit>=1.42.2",
    "twilio>=9.4.6",
    "tabulate>=0.9.0",
]

[tool.poetry.dependencies]
python-dotenv = "^1.0.0"
